app/src/main/python/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
aiohttp==3.9.1
    # via
    #   langchain
    #   langchain-community
aiosignal==1.3.1
    # via aiohttp
anyio==4.2.0
    # via
    #   httpx
    #   langchain-core
    #   openai
attrs==23.2.0
    # via aiohttp
azure-ai-formrecognizer==3.3.3
    # via -r requirements.in
azure-common==1.1.28
    # via azure-ai-formrecognizer
azure-core==1.29.7
    # via
    #   -r requirements.in
    #   azure-ai-formrecognizer
    #   msrest
certifi==2023.11.17
    # via
    #   httpcore
    #   httpx
    #   msrest
    #   requests
charset-normalizer==3.3.2
    # via requests
colorama==0.4.6
    # via tqdm
dataclasses-json==0.6.3
    # via
    #   langchain
    #   langchain-community
distro==1.9.0
    # via openai
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
greenlet==3.0.3
    # via sqlalchemy
h11==0.14.0
    # via httpcore
httpcore==1.0.2
    # via httpx
httpx==0.26.0
    # via openai
idna==3.6
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
isodate==0.6.1
    # via msrest
jsonpatch==1.33
    # via
    #   langchain
    #   langchain-core
jsonpointer==2.4
    # via jsonpatch
langchain==0.1.0
    # via -r requirements.in
langchain-community==0.0.11
    # via langchain
langchain-core==0.1.8
    # via
    #   langchain
    #   langchain-community
langsmith==0.0.83
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
marshmallow==3.20.2
    # via dataclasses-json
msrest==0.7.1
    # via azure-ai-formrecognizer
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via typing-inspect
numpy==1.26.3
    # via
    #   -r requirements.in
    #   langchain
    #   langchain-community
    #   opencv-python
    #   pandas
oauthlib==3.2.2
    # via requests-oauthlib
openai==1.9.0
    # via -r requirements.in
opencv-python==4.10.0.84
    # via -r requirements.in
packaging==23.2
    # via
    #   langchain-core
    #   marshmallow
    #   pytesseract
pandas==2.2.0
    # via -r requirements.in
pdf2image==1.17.0
    # via -r requirements.in
pillow==10.4.0
    # via
    #   -r requirements.in
    #   pdf2image
    #   pytesseract
pydantic==1.10.0
    # via
    #   -r requirements.in
    #   langchain
    #   langchain-core
    #   langsmith
    #   openai
pypdf==4.0.0
    # via -r requirements.in
pytesseract==0.3.10
    # via -r requirements.in
python-dateutil==2.8.2
    # via pandas
python-dotenv==1.0.1
    # via -r requirements.in
pytz==2023.3.post1
    # via pandas
pyyaml==6.0.1
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
regex==2023.12.25
    # via
    #   -r requirements.in
    #   tiktoken
requests==2.31.0
    # via
    #   azure-core
    #   langchain
    #   langchain-community
    #   langchain-core
    #   langsmith
    #   msrest
    #   requests-oauthlib
    #   tiktoken
requests-oauthlib==1.3.1
    # via msrest
six==1.16.0
    # via
    #   azure-core
    #   isodate
    #   python-dateutil
sniffio==1.3.0
    # via
    #   anyio
    #   httpx
    #   openai
sqlalchemy==2.0.25
    # via
    #   langchain
    #   langchain-community
tenacity==8.2.3
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
tesseract==0.1.3
    # via -r requirements.in
#tiktoken==0.5.2
    # via -r requirements.in
tqdm==4.66.1
    # via openai
typing-extensions==4.9.0
    # via
    #   -r requirements.in
    #   azure-ai-formrecognizer
    #   azure-core
    #   openai
    #   pydantic
    #   sqlalchemy
    #   typing-inspect
typing-inspect==0.9.0
    # via dataclasses-json
tzdata==2023.4
    # via pandas
urllib3==2.0.7
    # via requests
yarl==1.9.4
    # via aiohttp
